backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.7.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt==2.10.1
python-dotenv==1.0.1
pydantic[email]==2.10.3
celery==5.4.0
redis==5.2.1
docker==7.1.0
psycopg2-binary==2.9.10
